dynamic tiling to follow bottom shape
Tile to a reference dimension given by a second shape bottom. The `axis`
dimension of the first source bottom is tiled to equal the `axis`
dimension of the second shape bottom. For example, if the first bottom
dimension is 2 and the second bottom dimension is 6 the number of tiles
is set to 3. The reference dimension given by the shape bottom must be
divisible by the source dimension.

The static tiling specified by the `tiles` argument is unchanged.